Disable @covers check
We want to use PHPUnit 11 which no longer supports adding `@covers`
annotations and uses PHP attributes instead.

We're enforcing the existence of coverage information with the
`requireCoversAnnotation` config setting of PHPUnit.

This commit removes an optional feature, so we can do a new minor
release of the coding style.

Ticket: https://phabricator.wikimedia.org/T359971